requirements.txt
python-dotenv~=1.0.0
python-telegram-bot[webhooks]==20.2
requests